Move MigrationStoragePathSanitizer to demo module.
