+++ demo-simple/src/test/java/com/novoda/downloadmanager/demo/migration/MigrationStoragePathSanitizerTest.java
@@ -1,4 +1,4 @@
-package com.novoda.downloadmanager;
+package com.novoda.downloadmanager.demo.migration;
 
 import org.junit.Test;
